==> 01-clean-supp_data_4.ipynb <==
This notebook loosely follows 04_extract_data_from_supplementary_excel_files.<br>It is a pure data-cleaning notebook.

Input: NIHMS687993-supplement-supp_data_4.xlsx was downloaded from [here](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4481139/).<br>Output: NIHMS687993-supplement-supp_data_4v2.csv

import pandas as pd
from functions.data import (cluster_markers_raw_df_to_tables,
                            cluster_markers_tables_to_df,
                            format_cluster_markers_datatypes)

to_save=True

# Get Data

cluster_markers_raw_df = pd.read_excel('data/downloads/NIHMS687993-supplement-supp_data_4.xlsx',
                                       sheet_name='FINAL_MARKERS_FOR_EACH_CLUSTER.',
                                       header=None, names=['gene_symbol', 'myAUC', 'myDiff', 'power', 'cluster_no'])
cluster_markers_raw_df = cluster_markers_raw_df.rename(columns={"Unnamed: 0": 'gene_symbol'})
print(cluster_markers_raw_df.shape)
cluster_markers_raw_df

# Clean Data

cluster_markers_tables = cluster_markers_raw_df_to_tables(cluster_markers_raw_df)
cluster_markers_df = cluster_markers_tables_to_df(cluster_markers_tables)
cluster_markers_df = format_cluster_markers_datatypes(cluster_markers_df)
cluster_markers_df

if to_save:
    cluster_markers_df.to_csv('data/NIHMS687993-supplement-supp_data_4v2.csv', index=None)
