==> 02-make-mouse-gene-metadata.ipynb <==
This notebook loosely follows 04_extract_data_from_supplementary_excel_files.<br>It is a pure data-cleaning notebook.

Inputs:
1. NIHMS687993-supplement-supp_data_4v2.csv was prepared from 04-clean-supp_data_4
2. NIHMS687993-supplement-supp_data_2.xlsx was downloaded from [here](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4481139/).

Outputs:
1. mouse_gene_metadata.csv

import pandas as pd
pd.options.mode.chained_assignment = None

to_save=True

# Get Data

# FINAL_MARKERS_FOR_EACH_CLUSTER
cluster_markers_df = pd.read_csv('data/NIHMS687993-supplement-supp_data_4v2.csv')
cluster_markers_df['cluster_name'] = 'retina_' + cluster_markers_df['cluster_no'].astype(str).str.zfill(2)
cluster_markers_df['in_cluster'] = True
cluster_markers_df.head()

# Cell Cycle Regulated Genes
mmc2 = pd.read_excel('data/downloads/NIHMS687993-supplement-supp_data_2.xlsx',
                     sheet_name='Cell Cycle Regulated Genes',
                     names=['human_cluster', 'human_gene', 'cluster_no', 'gene_symbol',
                            None, 'all_mouse_genes', 'novel_mouse_genes', 'annotation']
                    )
mouse_genes = mmc2[['cluster_no', 'gene_symbol']]
mouse_genes['cluster_name'] = 'cellcycle_' + mouse_genes['cluster_no'].astype(str).str.zfill(2)
mouse_genes['in_cluster'] = True
mouse_genes[['gene_symbol', 'cluster_no', 'cluster_name', 'in_cluster']]

# Combine and Reshape

retina_metadata = cluster_markers_df.pivot(index='gene_symbol',
                                           columns='cluster_name',
                                           values='in_cluster')
retina_metadata = retina_metadata.fillna(False)
print(retina_metadata.shape)
retina_metadata.head()

cell_cycle_metadata = mouse_genes.pivot(index='gene_symbol', 
                              columns='cluster_name', 
                              values='in_cluster')
cell_cycle_metadata = cell_cycle_metadata.fillna(False)
print(cell_cycle_metadata.shape)
cell_cycle_metadata.head()

mouse_gene_metadata = pd.concat([retina_metadata, cell_cycle_metadata], axis=1)
mouse_gene_metadata = mouse_gene_metadata.fillna(False)
mouse_gene_metadata
print(mouse_gene_metadata.shape)
mouse_gene_metadata.head()

if to_save:
    mouse_gene_metadata.to_csv('data/mouse_gene_metadata.csv', index=True)
