==> 06-pca-on-differential_clusters.ipynb <==
This notebook follows 05_make_rentina_subsets_for_teaching.

Input: differential_clusters_expression.csv

Output:
1. differential_clusters_lowrank.csv
2. differential_clusters_sparse.csv

import os
import pandas as pd
import r_pca

to_save=True

# Get Data

retina_clusters_expression_subset = pd.read_csv('data/retina/differential_clusters_expression.csv',
                                                index_col=0)
retina_clusters_expression_subset[['NRXN2', 'ATP1B1', 'PAX6', 'SLC32A1', 'SLC6A1', 'ELAVL3']]

# PCA

# Warning: this takes a long time...
smusher = r_pca.R_pca(retina_clusters_expression_subset.to_numpy())
%time lowrank, sparse = smusher.fit(max_iter=200)  # maybe change this to 105 to match Olga's notebook?
lowrank

# Lowrank

lowrank_df = pd.DataFrame(lowrank,
                          index=retina_clusters_expression_subset.index,
                          columns=retina_clusters_expression_subset.columns)
lowrank_df.index.name = 'cell'
print(lowrank_df.shape)
lowrank_df.head()

if to_save:
    os.makedirs("data/pca", exist_ok=True)
    lowrank_df.to_csv('data/pca/differential_clusters_pca_lowrank.csv', index=True)

# Sparse

sparse_df = pd.DataFrame(sparse,
                         index=table1_diff_genes_cells.index,
                         columns=table1_diff_genes_cells.columns)
sparse_df.index.name = 'cell'
print(sparse_df.shape)
sparse_df.head()

if to_save:
    os.makedirs("data/pca", exist_ok=True)
    sparse_df.to_csv('data/pca/differential_clusters_pca_sparse.csv', index=True)
