==> 08-pca-and-tsne-on-differential_clusters.ipynb <==
This notebook follows 50_Example_workflow_reanalyzing_macosko2015.

Input: differential_clusters_expression.csv

import pandas as pd
import matplotlib.pyplot as plt
import phenograph

from sklearn.decomposition import PCA, NMF, FastICA
from sklearn.manifold import TSNE, MDS

retina_clusters_expression = pd.read_csv('data/retina/differential_clusters_expression.csv',
                                                index_col=0)
retina_clusters_expression[['NRXN2', 'ATP1B1', 'PAX6', 'SLC32A1', 'SLC6A1', 'ELAVL3']]

# PCA

# reduce the number of dimensions for t-SNE
data = retina_clusters_expression

smusher = PCA(n_components=15)
pca_array = smusher.fit_transform(data)
pca_df = pd.DataFrame(pca_array, index=data.index)

pca_df

# t-SNE Clustering on PCA

%%time

smusher = TSNE()
tsne_array = smusher.fit_transform(pca_df)

tsne_array
